PostModal: COmponent compounding
